refactor: extract shared messaging helper from popup and options
